requirements-airflow.txt
# Core Data & ML
pandas==2.3.3
scikit-learn==1.7.2
numpy==2.3.5
pyarrow==22.0.0

# Database & ORM (CRITICAL FIX)
# Airflow creates a crash loop with SQLAlchemy 2.0+ in this configuration.
sqlalchemy<2.0

# MLOps & Tracking
mlflow==3.7.0
boto3==1.42.4
skl2onnx==1.18.0
onnx==1.17.0
onnxruntime==1.20.1


# Data Validation
great_expectations==1.9.2

# Streaming (Lab 6)
confluent-kafka==2.12.2

# Configuration & Utilities
pyyaml==6.0.3

# Database Drivers (Required for Airflow Postgres Backend)
psycopg2-binary==2.9.10
